Prevent delete own account

+++ src/main/java/io/muic/ooc/webapp/servlet/DeleteUserServlet.java
@@ -9,6 +9,7 @@ import io.muic.ooc.webapp.Routable;
 import io.muic.ooc.webapp.model.User;
 import io.muic.ooc.webapp.service.SecurityService;
 import io.muic.ooc.webapp.service.UserService;
+import org.apache.commons.lang.StringUtils;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -44,17 +45,23 @@ public class DeleteUserServlet extends HttpServlet implements Routable {
 
             try{
                 User currentUser = userService.findByUsername(username);
+
                 User deletingUser = userService.findByUsername(request.getParameter("username"));
 
-                if (userService.deleteUserByUsername(deletingUser.getUsername())){
-                    request.getSession().setAttribute("hasError", false);
-                    request.getSession().setAttribute("message", String.format("User %s is successfully deleted.", deletingUser.getUsername()));
-
-                }else{
+                if(StringUtils.equals(currentUser.getUsername(), deletingUser.getUsername())){
                     request.getSession().setAttribute("hasError", true);
-                    request.getSession().setAttribute("message", String.format("Unable to delete user %s.", deletingUser.getUsername()));
-                }
+                    request.getSession().setAttribute("message", "You can not delete your own account.");
+                }else{
+                    if (userService.deleteUserByUsername(deletingUser.getUsername())){
+                        request.getSession().setAttribute("hasError", false);
+                        request.getSession().setAttribute("message", String.format("User %s is successfully deleted.", deletingUser.getUsername()));
 
+                    }else{
+                        request.getSession().setAttribute("hasError", true);
+                        request.getSession().setAttribute("message", String.format("Unable to delete user %s.", deletingUser.getUsername()));
+                    }
+
+                }
             }catch (Exception e){
                 request.getSession().setAttribute("hasError", true);
                 request.getSession().setAttribute("message", String.format("Unable to delete user %s.", request.getParameter("username")));
